+++ react/SearchWrapper.tsx
@@ -287,6 +287,7 @@ interface SearchWrapperProps {
   orderBy?: string
   to?: number
   page?: number
+  maxItemsPerPage?: number
   CustomContext?: ComponentType
   facetsLoading?: boolean
 }
@@ -307,6 +308,7 @@ const SearchWrapper: FC<SearchWrapperProps> = props => {
     CustomContext,
     page: pageFromQuery = 0,
     to,
+    maxItemsPerPage,
     children,
   } = props
   const {
@@ -380,7 +382,9 @@ const SearchWrapper: FC<SearchWrapperProps> = props => {
       department: searchQuery?.data
         ? getDepartmentMetadata(searchQuery.data)
         : null,
-      search: searchQuery?.data ? getSearchMetadata(searchQuery.data) : null,
+      search: searchQuery?.data
+        ? getSearchMetadata(searchQuery.data, pageFromQuery, maxItemsPerPage)
+        : null,
     }
 
     return [
@@ -390,7 +394,18 @@ const SearchWrapper: FC<SearchWrapperProps> = props => {
         products,
       },
     ]
-  }, [searchQuery, facetsLoading, params, account, title, orderBy, map, page])
+  }, [
+    searchQuery,
+    facetsLoading,
+    params,
+    account,
+    title,
+    orderBy,
+    map,
+    page,
+    pageFromQuery,
+    maxItemsPerPage,
+  ])
 
   const [hasLoaded, setHasLoaded] = useState(true)