+++ react/modules/searchMetadata.ts
@@ -103,7 +103,11 @@ export const getCategoryMetadata = (searchQuery?: SearchQueryData) => {
   }
 }
 
-export const getSearchMetadata = (searchQuery?: SearchQueryData) => {
+export const getSearchMetadata = (
+  searchQuery?: SearchQueryData,
+  page?: number,
+  maxItemsPerPage?: number
+) => {
   if (
     !searchQuery ||
     !searchQuery.productSearch ||
@@ -128,6 +132,12 @@ export const getSearchMetadata = (searchQuery?: SearchQueryData) => {
 
   const department = getDepartment(searchQuery)
 
+  const products = searchQuery?.products?.map(product => {
+    return {
+      productId: product.productId,
+    }
+  })
+
   return {
     term: decodedTerm || undefined,
     category: department ? { id: department.id, name: department.name } : null,
@@ -135,5 +145,8 @@ export const getSearchMetadata = (searchQuery?: SearchQueryData) => {
     operator: searchQuery.productSearch.operator,
     searchState: searchQuery.productSearch.searchState,
     correction: searchQuery.productSearch.correction,
+    products,
+    page,
+    maxItemsPerPage,
   }
 }